backend1/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
blis==1.0.1
boto3==1.35.73
botocore==1.35.73
camelot-py==0.9.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
cryptography==44.0.0
cymem==2.0.10
dataclasses-json==0.6.7
distro==1.9.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
et_xmlfile==2.0.0
exceptiongroup==1.2.2
faiss-cpu==1.9.0.post1
fastapi==0.115.5
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
httpx-sse==0.4.0
huggingface-hub==0.26.3
idna==3.10
interchange==2021.0.4
Jinja2==3.1.4
jiter==0.8.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.9
langchain-community==0.3.9
langchain-core==0.3.21
langchain-neo4j==0.1.1
langchain-openai==0.2.11
langchain-text-splitters==0.3.2
langcodes==3.5.0
langsmith==0.1.147
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.11
mypy-extensions==1.0.0
neo4j==5.27.0
networkx==3.2.1
numpy==1.26.4
openai==1.56.0
opencv-python==4.10.0.84
openpyxl==3.1.5
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pansi==2024.11.0
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
preshed==3.0.9
propcache==0.2.1
psutil==6.1.0
py2neo==2021.2.4
pycparser==2.22
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
Pygments==2.18.0
PyMuPDF==1.24.14
pypdf==5.1.0
PyPDF2==2.12.1
pypdfium2==4.30.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
s3transfer==0.10.4
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
sentence-transformers==3.3.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
srsly==2.4.8
starlette==0.41.3
sympy==1.13.1
tabula-py==2.10.0
tabulate==0.9.0
tenacity==9.0.0
thinc==8.3.2
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.1
transformers==4.46.3
typer==0.14.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
uvicorn==0.32.1
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.0
yarl==1.18.3
